Change the type of reserved memory from int to float

+++ src/Frame/Frame.h
@@ -85,7 +85,7 @@ class Frame {
 public:
     // Load image cache for default_z, except for PV preview image which needs cube
     Frame(uint32_t session_id, std::shared_ptr<FileLoader> loader, const std::string& hdu, int default_z = DEFAULT_Z,
-        int reserved_memory = 0);
+        float reserved_memory = 0);
     ~Frame(){};
 
     bool IsValid();
@@ -221,7 +221,7 @@ public:
         std::vector<float>& data, int& downsampled_width, int& downsampled_height, int z, int stokes, CARTA::ImageBounds& bounds, int mip);
     bool CalculateVectorField(const std::function<void(CARTA::VectorOverlayTileData&)>& callback);
 
-    int UsedReservedMemory() const;
+    float UsedReservedMemory() const;
     bool GetPointSpectralData(std::vector<float>& profile, int stokes, PointXy point);
     bool GetRegionSpectralData(const AxisRange& z_range, int stokes, const casacore::ArrayLattice<casacore::Bool>& mask,
         const casacore::IPosition& origin, std::map<CARTA::StatsType, std::vector<double>>& profiles);
